moved filtering to separate file
